+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ------
 
+* [#1221](https://github.com/Shopify/shopify-app-cli/pull/1221): Prioritizes returning an HTTPS URL over HTTP from `shopify tunnel status`.
+
 Version 1.10.0
 -------------
 * Updating internal features in development